Reverted config eject
