--- config/jest/typescriptTransform.js
@@ -1,7 +1,0 @@
-// Copyright 2004-present Facebook. All Rights Reserved.
-
-'use strict';
-
-const tsJestPreprocessor = require('ts-jest/preprocessor');
-
-module.exports = tsJestPreprocessor;
