+++ src/App.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 
 import Bar from './components/Bar';
-import SendEmailButton from './components/SendEmailButton';
 import Steps from './components/Stepper/Steps';
 import { UserCtx } from './contexts/user.context';
 import { useLocalStorage } from './hooks/localstorage.hook';
