+++ src/components/Auth/Login.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import GoogleLogin from 'react-google-login';
 import { UserCtx } from "../../contexts/user.context";
-import { createUserFromJson, IUser } from '../../types/user';
+import { createUserFromJson, IUser } from '../../types';
 
 const CLIENT_ID = process.env.REACT_APP_GOOGLE_ID || '';
