+++ src/components/Auth/LogoutButton.tsx
@@ -1,12 +1,12 @@
 import { Button } from '@material-ui/core';
 import React, { useContext } from 'react';
-import { UserCtx } from 'src/contexts/user.context';
+import {UserCtx} from "../../contexts/user.context";
 
-const logoutButton = () => {
+const LogoutButton = () => {
   const [, , removeUser] = useContext(UserCtx);
   return (
     <Button variant="outlined" color="inherit" onClick={removeUser}>Log out</Button>
   )
-}
+};
 
-export default logoutButton;
+export default LogoutButton;
