+++ src/components/Bar.tsx
@@ -1,13 +1,11 @@
-import PropTypes from 'prop-types';
-import React from 'react';
-
 import AppBar from '@material-ui/core/AppBar';
 import { createStyles, WithStyles, withStyles } from '@material-ui/core/styles';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
+import PropTypes from 'prop-types';
+import React from 'react';
 import LogoutButton from '../components/Auth/LogoutButton';
 import { UserCtx } from '../contexts/user.context';
-
 import Login from './Auth/Login';
 import Hello from './Hello';
