+++ src/components/Editor/DynamicVariables.tsx
@@ -2,7 +2,7 @@ import Chip from '@material-ui/core/Chip';
 import Paper from '@material-ui/core/Paper';
 import Typography from '@material-ui/core/Typography';
 import React from 'react';
-import { SpreadsheetCtx } from 'src/contexts/spreadsheet.context';
+import {SpreadsheetCtx} from "../../contexts/spreadsheet.context";
 
 const DynamicVariables = () => {
   const [spreadsheet] = React.useContext(SpreadsheetCtx);
