+++ src/components/Editor/Editor.tsx
@@ -1,14 +1,12 @@
 import Grid from '@material-ui/core/Grid';
 import { EditorState } from 'draft-js';
-import { Editor } from 'react-draft-wysiwyg';
-import { SpreadsheetCtx } from 'src/contexts/spreadsheet.context';
-import { mailContent } from 'src/seeds/mail';
-
-import React from 'react';
-import DynamicVariables from './DynamicVariables';
-
 import { stateToHTML } from 'draft-js-export-html';
 import { stateFromHTML } from 'draft-js-import-html';
+import React from 'react';
+import { Editor as Wysiwyg } from 'react-draft-wysiwyg';
+import {SpreadsheetCtx} from "../../contexts/spreadsheet.context";
+import {mailContent} from "../../seeds/mail";
+import DynamicVariables from './DynamicVariables';
 
 import 'react-draft-wysiwyg/dist/react-draft-wysiwyg.css';
 
@@ -21,7 +19,7 @@ const replaceVars = (input: string, spreadsheet: any) => {
   });
 };
 
-const editor = () => {
+const Editor = () => {
   const [spreadsheet, setSpreadsheet] = React.useContext(SpreadsheetCtx);
   const [editor, setEditor] = React.useState(
     EditorState.createWithContent(stateFromHTML(mailContent)),
@@ -41,7 +39,7 @@ const editor = () => {
     <div style={{ padding: 20 }}>
       <Grid>
         <DynamicVariables />
-        <Editor
+        <Wysiwyg
           editorState={editor}
           onEditorStateChange={onChange}
           toolbar={{
@@ -55,4 +53,4 @@ const editor = () => {
   );
 };
 
-export default editor;
+export default Editor;
