+++ src/components/Sender.tsx
@@ -1,18 +1,16 @@
-import { Button, Grid } from '@material-ui/core';
-import SendEmailButton from 'src/components/SendEmailButton';
-
+import { Grid } from '@material-ui/core';
 import React from 'react';
-import { MailTemplateCtx } from 'src/contexts/mail-template.context';
-import { SpreadsheetCtx } from 'src/contexts/spreadsheet.context';
+import {MailTemplateCtx} from "../contexts/mail-template.context";
+import {SpreadsheetCtx} from "../contexts/spreadsheet.context";
+import {IUser} from "../types";
+import SendEmailButton from "./SendEmailButton";
 import { replaceVars } from './utils';
 
 const Sender = () => {
-  const [mailTemplate] = React.useContext(MailTemplateCtx);
+  const mailTemplate = React.useContext(MailTemplateCtx);
   const [spreadsheet] = React.useContext(SpreadsheetCtx);
-  // console.log(spreadsheet);
-  // console.log(mailTemplate);
 
-  const dataToSend = spreadsheet.usersData.map(user => {
+  const dataToSend = spreadsheet.usersData.map((user: IUser) => {
     return {
       email: user.email,
       data: {
