+++ src/components/Stepper/Steps.tsx
@@ -2,13 +2,12 @@ import Step from '@material-ui/core/Step';
 import StepButton from '@material-ui/core/StepButton';
 import Stepper from '@material-ui/core/Stepper';
 import React from 'react';
-import { SpreadsheetCtx } from 'src/contexts/spreadsheet.context';
-import { StepCtx } from 'src/contexts/step.context';
-import { IStep } from 'src/types/step';
+import {getStep} from "../../const/steps";
 import { MailTemplateCtx } from '../../contexts/mail-template.context';
+import {SpreadsheetCtx} from '../../contexts/spreadsheet.context';
+import {StepCtx} from '../../contexts/step.context';
 import { mailContent } from '../../seeds/mail';
-
-import { getStep } from 'src/const/steps';
+import {IStep} from "../../types";
 import Navigation from './Navigation';
 
 const Steps = () => {
