+++ src/contexts/mail-template.context.tsx
@@ -1,4 +1,4 @@
 import { createContext } from 'react';
 import { mailContent } from '../seeds/mail'
 
-export const MailTemplateCtx = createContext(mailContent);
+export const MailTemplateCtx = createContext<any>(mailContent);
