+++ src/services/spreadsheet.service.ts
@@ -1,6 +1,6 @@
-import { SpreadSheetProvider } from 'src/providers/spreadsheet.provider';
-import { SpreadSheetTransofrmer } from 'src/transformers/spreadsheet.transformer';
-import { ISpreadsheet } from 'src/types';
+import {SpreadSheetProvider} from "../providers/spreadsheet.provider";
+import {SpreadSheetTransofrmer} from "../transformers/spreadsheet.transformer";
+import {ISpreadsheet} from "../types";
 
 export default class SpreadSheetService {
   public onFilePicked(data: any, setSpreadsheet: (data: ISpreadsheet) => void) {
